created customHook to check if screen is mobile or not
